Insert status in GitHub issue indexing reports
First for convenience: you can see what's happened with just a look at
the email subject or issue title.

Second, to solve a problem with the current way we add comments:
we never add a comment for successes, and we don't duplicate warning
comments... so when reindexing occurs after a warning, we don't know
whether the new indexing has succeeded or still has the same warnings!