Do not remove stopwords from search
First because it's not strictly necessary, as the BM25 scoring we use
(the default) takes inverse document frequency into account, so that a
stopword is much less likely to impact the score.

Second, because we require all terms given by the user to match, so
stopwords won't add a gazillion more matching documents.

And third, because removing stopwords leads to strange behavior with
the fast-vector highlighter.

Note I kept the cn/ja stopwords removers, because those seem to do more
than remove stopwords. Without them, I got highlighting on commas, for
example...